requirements.txt
folium==0.19.5
geopy==2.4.1
joblib==1.4.0
openrouteservice==2.3.3
pandas==2.2.3
prophet==1.1.6
python-dotenv==1.1.0
Requests==2.32.3
scikit_learn==1.4.2
streamlit_folium==0.23.1
streamlit_navigation_bar==3.3.0
streamlit==1.10.0


